Extract metrics.go from server.go
