+++ pkg/server/server.go
@@ -44,20 +44,6 @@ var (
 	plog = capnslog.NewPackageLogger("github.com/openshift/console", "server")
 )
 
-var (
-	appVersion = prometheus.NewGauge(prometheus.GaugeOpts{
-		Name: "console_version",
-		Help: "Version information about this binary",
-		ConstLabels: map[string]string{
-			"version": "v0.1.0",
-		},
-	})
-	apiRequestsTotal = prometheus.NewCounterVec(prometheus.CounterOpts{
-		Name: "console_api_requests_total",
-		Help: "Count of all HTTP requests against the API",
-	}, []string{"code", "method", "path"})
-)
-
 type jsGlobals struct {
 	ConsoleVersion           string `json:"consoleVersion"`
 	AuthDisabled             bool   `json:"authDisabled"`
@@ -126,29 +112,6 @@ func (s *Server) alertManagerProxyEnabled() bool {
 func (s *Server) meteringProxyEnabled() bool {
 	return s.MeteringProxyConfig != nil
 }
-
-// NOTE: by default, promhttp provides:
-//   promhttp.InstrumentHandlerCounter(apiRequestsTotal, handler)
-// but this handler does not allow for the additional "path" label, which
-// we desire.
-func InstrumentAPIHandlerCounter(counter *prometheus.CounterVec, next http.Handler, path string) http.HandlerFunc {
-	return http.HandlerFunc(func(writer http.ResponseWriter, req *http.Request) {
-
-		// we need to let the request be served before we can check the response status
-		next.ServeHTTP(writer, req)
-
-		apiRequestsTotal.With(prometheus.Labels{
-			// TODO: this is a little fiddly to unwrap
-			"code": "TODO",
-			"method": req.Method,
-			"path": path,
-		}).Inc()
-	})
-}
-
-// to use for page view when we track those
-// req.Header.Get("User-Agent")
-func InstrumentPageViewCounter() {}
 
 func (s *Server) HTTPHandler() http.Handler {
 	// TODO: decide where to place this.
